Reestructuracion de Adopcion

+++ backend/routes/adopciones.php
@@ -1,28 +1,60 @@
 <?php
-require_once "../controllers/AdopcionController.php";
+
+require_once __DIR__ . "/../controllers/AdopcionController.php";
 
 $controller = new AdopcionController();
-$data = json_decode(file_get_contents("php://input"), true);
+
+$method = $_SERVER['REQUEST_METHOD'];
+
+$id = $_GET['id'] ?? null;
+
+$data = json_decode(
+    file_get_contents("php://input"),
+    true
+);
 
 switch ($method) {
 
     case 'GET':
+
         if ($id) {
             $controller->show($id);
         } else {
             $controller->index();
         }
+
         break;
 
     case 'POST':
+
         $controller->store($data);
+
         break;
 
     case 'PUT':
-        $controller->update($id, $data);
+
+        if ($id) {
+            $controller->update($id, $data);
+        }
+
         break;
 
     case 'DELETE':
-        $controller->delete($id);
+
+        if ($id) {
+            $controller->delete($id);
+        }
+
+        break;
+
+    default:
+
+        http_response_code(405);
+
+        echo json_encode([
+            "success" => false,
+            "message" => "Método no permitido"
+        ]);
+
         break;
 }
